+++ packages/client/src/App.tsx
@@ -6,6 +6,7 @@
 
 import { Routes, Route } from "react-router";
 import { useSocket } from "./hooks/useSocket.js";
+import { Toast } from "./components/Toast.js";
 import { TopPage } from "./pages/TopPage.js";
 import { CreateRoomPage } from "./pages/CreateRoomPage.js";
 import { JoinRoomPage } from "./pages/JoinRoomPage.js";
@@ -16,12 +17,15 @@ export function App() {
   useSocket();
 
   return (
-    <Routes>
-      <Route path="/" element={<TopPage />} />
-      <Route path="/create" element={<CreateRoomPage />} />
-      <Route path="/join" element={<JoinRoomPage />} />
-      <Route path="/join/:roomCode" element={<JoinRoomPage />} />
-      <Route path="/room/:roomCode" element={<RoomPage />} />
-    </Routes>
+    <>
+      <Toast />
+      <Routes>
+        <Route path="/" element={<TopPage />} />
+        <Route path="/create" element={<CreateRoomPage />} />
+        <Route path="/join" element={<JoinRoomPage />} />
+        <Route path="/join/:roomCode" element={<JoinRoomPage />} />
+        <Route path="/room/:roomCode" element={<RoomPage />} />
+      </Routes>
+    </>
   );
 }